Philly Junkyard Locations and Business Licenses
All past and current Philly business licenses for locations that have ever had a junkyard license, with parcel data attached. The first column in the dataframe is the official address of the property. The geometry column is shapes/polygons of the parcel. The original data was points, which are preserved in the geocode_x, geocode_y columns in case we need them for some reason. The columns are a mixture of columns from the L&I Business Licenses data and Philly parcel data (see notebooks for links and details).
To import to geopandas: df = gpd.read_file("PhillyJunkLicensesLocations.geojson")