Philly Junkyard Locations and Basic Info
This dataset contains only the junkyard licenses (past and current) and locations, not all other licenses attached to those sites. It does have details on owners, status of the junkyard license, and other business license categories. Like the business license dataset, this has been merged with parcel data to make the geometry polygons/shapes. Original points are preserved in the geocode_x, geocode_y columns. It is in CRS 2272. "parcel_full" is probably the best index/merge key to use.